feat: merge wasm/vexos panic impls
